+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### ✨ Improved
 
 * Monitor lock file and abort if it is deleted during operation.
+* Add `clear-lock` command to remove existing lock file.
 * Ensure all valves are closed at the end of the LN2 fill process.
 * Do not run post-fill validation if the fill was aborted or failed.
 * Add retrier for `LN2Handler.stop()`.